added admin api routes, admin role
